Refactored spec test
