+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/application/client.stream.rpc/message.exchange/client.rpt
@@ -44,6 +44,4 @@ read ${grpc:protobuf()
            .string(1, "Hello World")
            .build()}
 
-write close
-
 read closed
