+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/application/client.stream.rpc/message.exchange/server.rpt
@@ -43,6 +43,4 @@ write ${grpc:protobuf()
             .build()}
 write flush
 
-read closed
-
 write close
