+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/application/server.stream.rpc/message.exchange/client.rpt
@@ -45,4 +45,3 @@ read ${grpc:protobuf()
            .string(1, "Hello World2")
            .build()}
 
-read closed
