+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/application/server.stream.rpc/message.exchange/server.rpt
@@ -45,5 +45,3 @@ write ${grpc:protobuf()
             .string(1, "Hello World2")
             .build()}
 write flush
-
-write close
