--- incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/application/unary.rpc/server.rpt
@@ -1,44 +1,0 @@
-#
-# Copyright 2021-2022 Aklivity Inc
-#
-# Licensed under the Aklivity Community License (the "License"); you may not use
-# this file except in compliance with the License.  You may obtain a copy of the
-# License at
-#
-#   https://www.aklivity.io/aklivity-community-license/
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OF ANY KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations under the License.
-#
-
-accept "zilla://streams/app0"
-       option zilla:window 8192
-       option zilla:transmission "half-duplex"
-accepted
-
-read zilla:begin.ext ${grpc:matchBeginEx()
-                           .typeId(zilla:id("grpc"))
-                           .scheme("http")
-                           .authority("localhost:8080")
-                           .service("example.EchoService")
-                           .method("EchoUnary")
-                           .request("UNARY")
-                           .response("UNARY")
-                           .metadata("custom", "test")
-                           .build()}
-connected
-
-read ${grpc:protobuf()
-           .string(1, "Hello World")
-           .build()}
-
-read closed
-
-write ${grpc:protobuf()
-            .string(1, "Hello World")
-            .build()}
-write flush
-
-write close
