+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/network/bidirectional.stream.rpc/message.exchange/client.rpt
@@ -24,7 +24,6 @@ write zilla:begin.ext ${http:beginEx()
                             .header(":authority", "localhost:8080")
                             .header(":path", "/example.EchoService/EchoStream")
                             .header("content-type", "application/grpc")
-                            .header("te", "trailers")
                             .header("custom", "test")
                             .build()}
