+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/network/bidirectional.stream.rpc/message.exchange/server.rpt
@@ -25,7 +25,6 @@ read zilla:begin.ext ${http:matchBeginEx()
                              .header(":authority", "localhost:8080")
                              .header(":path", "/example.EchoService/EchoStream")
                              .header("content-type", "application/grpc")
-                             .header("te", "trailers")
                              .header("custom", "test")
                              .build()}
