+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/network/client.stream.rpc/message.exchange/client.rpt
@@ -24,7 +24,6 @@ write zilla:begin.ext ${http:beginEx()
                             .header(":authority", "localhost:8080")
                             .header(":path", "/example.EchoService/EchoClientStream")
                             .header("content-type", "application/grpc")
-                            .header("te", "trailers")
                             .header("custom", "test")
                             .build()}
 
@@ -51,7 +50,6 @@ read ${grpc:message()
            .string(1, "Hello World")
            .build()}
 
-write close
 
 read zilla:end.ext ${http:endEx()
                          .typeId(zilla:id("http"))
