+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/network/rejected.rpc/method.not.post/client.rpt
@@ -30,10 +30,19 @@ write zilla:begin.ext ${http:beginEx()
 
 connected
 
-read zilla:reset.ext ${http:resetEx()
+write ${grpc:message()
+            .string(1, "Hello World")
+            .build()}
+write flush
+
+write close
+
+read zilla:begin.ext ${http:matchBeginEx()
                            .typeId(zilla:id("http"))
                            .header(":status", "405")
+                           .header("content-type", "text/plain; encoding=utf-8")
                            .header("grpc-status", "13")
                            .build()}
 
-write aborted
+
+read closed
