+++ incubator/binding-grpc.spec/src/main/scripts/io/aklivity/zilla/specs/binding/grpc/streams/network/server.stream.rpc/message.exchange/server.rpt
@@ -25,7 +25,6 @@ read zilla:begin.ext ${http:matchBeginEx()
                              .header(":authority", "localhost:8080")
                              .header(":path", "/example.EchoService/EchoServerStream")
                              .header("content-type", "application/grpc")
-                             .header("te", "trailers")
                              .header("custom", "test")
                              .build()}
 
@@ -54,10 +53,3 @@ write ${grpc:message()
             .string(1, "Hello World2")
             .build()}
 write flush
-
-write zilla:end.ext ${http:endEx()
-                         .typeId(zilla:id("http"))
-                         .trailer("grpc-status", "0")
-                         .build()}
-
-write close
