+++ incubator/binding-grpc.spec/src/test/java/io/aklivity/zilla/specs/binding/grpc/streams/network/BidirectionalStreamRpcIT.java
@@ -12,7 +12,7 @@
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.aklivity.zilla.specs.binding.grpc.streams.application;
+package io.aklivity.zilla.specs.binding.grpc.streams.network;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
@@ -25,10 +25,12 @@ import org.junit.rules.Timeout;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
 
-public class StreamIT
+
+public class BidirectionalStreamRpcIT
 {
     private final K3poRule k3po = new K3poRule()
-        .addScriptRoot("app", "io/aklivity/zilla/specs/binding/grpc/streams/application");
+        .addScriptRoot("net",
+            "io/aklivity/zilla/specs/binding/grpc/streams/network/bidirectional.stream.rpc");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(10, SECONDS));
 
@@ -37,38 +39,8 @@ public class StreamIT
 
     @Test
     @Specification({
-        "${app}/unary.rpc/client",
-        "${app}/unary.rpc/server",
-    })
-    public void shouldEstablishUnaryRpc() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${app}/client.stream.rpc/client",
-        "${app}/client.stream.rpc/server",
-    })
-    public void shouldEstablishClientStreamRpc() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${app}/server.stream.rpc/client",
-        "${app}/server.stream.rpc/server",
-    })
-    public void shouldEstablishServerStreamRpc() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${app}/bidirectional.stream.rpc/client",
-        "${app}/bidirectional.stream.rpc/server",
+        "${net}/message.exchange/client",
+        "${net}/message.exchange/server",
     })
     public void shouldEstablishBidirectionalRpc() throws Exception
     {
