--- incubator/binding-grpc.spec/src/test/java/io/aklivity/zilla/specs/binding/grpc/streams/network/StreamIT.java
@@ -1,97 +1,0 @@
-/*
- * Copyright 2021-2022 Aklivity Inc
- *
- * Licensed under the Aklivity Community License (the "License"); you may not use
- * this file except in compliance with the License.  You may obtain a copy of the
- * License at
- *
- *   https://www.aklivity.io/aklivity-community-license/
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package io.aklivity.zilla.specs.binding.grpc.streams.network;
-
-import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.junit.rules.RuleChain.outerRule;
-
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.DisableOnDebug;
-import org.junit.rules.TestRule;
-import org.junit.rules.Timeout;
-import org.kaazing.k3po.junit.annotation.Specification;
-import org.kaazing.k3po.junit.rules.K3poRule;
-
-public class StreamIT
-{
-    private final K3poRule k3po = new K3poRule()
-        .addScriptRoot("net", "io/aklivity/zilla/specs/binding/grpc/streams/network");
-
-    private final TestRule timeout = new DisableOnDebug(new Timeout(10, SECONDS));
-
-    @Rule
-    public final TestRule chain = outerRule(k3po).around(timeout);
-
-    @Test
-    @Specification({
-        "${net}/unary.rpc/client",
-        "${net}/unary.rpc/server",
-    })
-    public void shouldEstablishUnaryRpc() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/client.stream.rpc/client",
-        "${net}/client.stream.rpc/server",
-    })
-    public void shouldEstablishClientStreamRpc() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/server.stream.rpc/client",
-        "${net}/server.stream.rpc/server",
-    })
-    public void shouldEstablishServerStreamRpc() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/bidirectional.stream.rpc/client",
-        "${net}/bidirectional.stream.rpc/server",
-    })
-    public void shouldEstablishBidirectionalRpc() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/method.not.post/client",
-        "${net}/method.not.post/server",
-    })
-    public void shouldRejectMethodNotPost() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/unrecognized.rpc/client",
-        "${net}/unrecognized.rpc/server",
-    })
-    public void shouldRejectUnrecognizedMethod() throws Exception
-    {
-        k3po.finish();
-    }
-}
