+++ incubator/binding-grpc.spec/src/test/java/io/aklivity/zilla/specs/binding/grpc/streams/network/UnaryRpcIT.java
@@ -45,64 +45,4 @@ public class UnaryRpcIT
     {
         k3po.finish();
     }
-
-    @Test
-    @Specification({
-        "${net}/binary.metadata/client",
-        "${net}/binary.metadata/server",
-    })
-    public void shouldEstablishWithBinaryMetadata() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/grpc.web/client",
-        "${net}/grpc.web/server",
-    })
-    public void shouldEstablishGrpcWeb() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/response.timeout/client",
-        "${net}/response.timeout/server",
-    })
-    public void shouldTimeoutOnNoResponse() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/server.send.read.abort.on.open.request/client",
-        "${net}/server.send.read.abort.on.open.request/server"
-    })
-    public void serverSendsReadAbortOnOpenRequest() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/server.send.write.abort.on.open.response/client",
-        "${net}/server.send.write.abort.on.open.response/server"
-    })
-    public void serverSendsWriteAbortOnOpenResponse() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/server.send.write.abort.on.open.request.response/client",
-        "${net}/server.send.write.abort.on.open.request.response/server"
-    })
-    public void serverSendsWriteAbortOnOpenRequestResponse() throws Exception
-    {
-        k3po.finish();
-    }
 }
